+++ src/rezbuild/__init__.py
@@ -2,6 +2,7 @@
 from rezbuild.builder import CompileBuilder
 from rezbuild.builder import ExtractBuilder
 from rezbuild.builder import InstallBuilder
+from rezbuild.builder import MacOSBuilder
 from rezbuild.builder import MacOSDmgBuilder
 from rezbuild.builder import PythonBuilder
 from rezbuild.builder import PythonSourceBuilder
@@ -13,6 +14,7 @@ __all__ = [
     "CompileBuilder",
     "ExtractBuilder",
     "InstallBuilder",
+    "MacOSBuilder",
     "MacOSDmgBuilder",
     "PythonBuilder",
     "PythonSourceBuilder",